modif2

+++ ft_strlen.c
@@ -14,9 +14,9 @@
 #include <string.h>
 #include <stddef.h>
 
-int	ft_strlen(const char *s)
+size_t	ft_strlen(const char *s)
 {
-	int	length;
+	size_t	length;
 
 	length = 0;
 	while (s[length] != '\0')
